+++ src/wallets/hardware/ledger/errorHandler.js
@@ -17,7 +17,9 @@ const ERRORS = {
   'Invalid networkId signature returned.': 'ledgerError.invalidNetworkID',
   'EthAppPleaseEnableContractData: Please enable Contract data on the Ethereum app Settings':
     'ledgerError.enableContractData',
-  'No device selected.': 'ledgerError.noDeviceSelected'
+  'No device selected.': 'ledgerError.noDeviceSelected',
+  'Incompatible EIP155-based V 10 and chain id 1. See the second parameter of the Transaction constructor to set the chain id.':
+    'ledgerError.appAndNetworkMismatch'
 };
 const WARNING = {
   'Ledger device: Condition of use not satisfied (denied by the user?) (0x6985)':