+++ tests/unit/specs/HDWalletInterface.mnemonic.spec.js
@@ -1,7 +1,7 @@
 import HDWalletInterface from '@/wallets/HDWalletInterface';
 import * as HDKey from 'hdkey';
 import { hashPersonalMessage, ecsign, toBuffer } from 'ethereumjs-util';
-import ethTx from 'ethereumjs-tx';
+import { Transaction } from 'ethereumjs-tx';
 
 const bip39 = require('bip39');
 const ETH_PATH = "m/44'/60'/0'/0";
@@ -61,7 +61,7 @@ for (let i = 0; i < 3; i++)
       'mnemonic',
       () => {},
       tx => {
-        tx = new ethTx(tx);
+        tx = new Transaction(tx);
         return new Promise(resolve => {
           tx.sign(hdk.derive(ETH_PATH + '/' + i).privateKey);
           resolve(tx);
@@ -99,7 +99,7 @@ describe('HDWalletInterface Mnemonic', () => {
         accountList[i].address
       );
       const signedTx = await testWallets[i].signTransaction(txParams);
-      const _signedTx = new ethTx(txParams);
+      const _signedTx = new Transaction(txParams);
       _signedTx.sign(
         Buffer.from(accountList[i].privkey.replace('0x', ''), 'hex')
       );